Deal with replayed (rebased or rewound) git sha
If a push event from github would trigger an autodeployment, but an
AutoDeployment already exists, log a message and return without creating
a duplicate AutoDeployment.

The AutoDeployment model is gaurded by a foriegn key contraint on the
sha field so this bug fix simply prevents an exception from firing
and provides a more useful log message.

	modified:   push_event.rb